[ISSUE#4233] Move the capability of slaveActingMaster from container module to broker module
