+++ container/src/test/java/org/apache/rocketmq/container/BrokerContainerTest.java
@@ -76,13 +76,13 @@ public class BrokerContainerTest {
             new NettyClientConfig(),
             new MessageStoreConfig());
 
+        brokerController.getBrokerConfig().setEnableSlaveActingMaster(true);
+
         BrokerOuterAPI brokerOuterAPI = mock(BrokerOuterAPI.class);
         Field field = BrokerController.class.getDeclaredField("brokerOuterAPI");
         field.setAccessible(true);
         field.set(brokerController, brokerOuterAPI);
 
-        // topic-0 doesn't have queueGroupConfig.
-        // topic-1 has queueGroupConfig.
         List<TopicConfig> topicConfigList = new ArrayList<>(2);
         for (int i = 0; i < 2; i++) {
             topicConfigList.add(new TopicConfig("topic-" + i));
@@ -108,13 +108,13 @@ public class BrokerContainerTest {
             new NettyClientConfig(),
             new MessageStoreConfig());
 
+        brokerController.getBrokerConfig().setEnableSlaveActingMaster(true);
+
         BrokerOuterAPI brokerOuterAPI = mock(BrokerOuterAPI.class);
         Field field = BrokerController.class.getDeclaredField("brokerOuterAPI");
         field.setAccessible(true);
         field.set(brokerController, brokerOuterAPI);
 
-        // topic-0 doesn't have queueGroupConfig.
-        // topic-1 has queueGroupConfig.
         List<TopicConfig> topicConfigList = new ArrayList<>(2);
         for (int i = 0; i < 2; i++) {
             topicConfigList.add(new TopicConfig("topic-" + i));
@@ -128,7 +128,7 @@ public class BrokerContainerTest {
         ArgumentCaptor<TopicConfigSerializeWrapper> captor = ArgumentCaptor.forClass(TopicConfigSerializeWrapper.class);
         ArgumentCaptor<BrokerIdentity> brokerIdentityCaptor = ArgumentCaptor.forClass(BrokerIdentity.class);
         verify(brokerOuterAPI).registerBrokerAll(anyString(), anyString(), anyString(), anyLong(), anyString(),
-            captor.capture(), ArgumentMatchers.anyList(), anyBoolean(), anyInt(), anyBoolean(), anyBoolean(), brokerIdentityCaptor.capture());
+            captor.capture(), ArgumentMatchers.anyList(), anyBoolean(), anyInt(), anyBoolean(), anyBoolean(), anyLong(), brokerIdentityCaptor.capture());
         TopicConfigSerializeWrapper wrapper = captor.getValue();
         for (Map.Entry<String, TopicConfig> entry : wrapper.getTopicConfigTable().entrySet()) {
             assertThat(entry.getValue().getPerm()).isEqualTo(brokerController.getBrokerConfig().getBrokerPermission());
@@ -245,9 +245,9 @@ public class BrokerContainerTest {
     @Test
     public void testAddAndRemoveDLedgerBroker() throws Exception {
         BrokerContainer brokerContainer = new BrokerContainer(
-                new BrokerContainerConfig(),
-                new NettyServerConfig(),
-                new NettyClientConfig());
+            new BrokerContainerConfig(),
+            new NettyServerConfig(),
+            new NettyClientConfig());
         assertThat(brokerContainer.initialize()).isTrue();
         brokerContainer.start();
 
@@ -354,7 +354,7 @@ public class BrokerContainerTest {
         ArgumentCaptor<TopicConfigSerializeWrapper> captor = ArgumentCaptor.forClass(TopicConfigSerializeWrapper.class);
         ArgumentCaptor<BrokerIdentity> brokerIdentityCaptor = ArgumentCaptor.forClass(BrokerIdentity.class);
         verify(brokerOuterAPI, times(times)).registerBrokerAll(anyString(), anyString(), anyString(), anyLong(),
-            anyString(), captor.capture(), ArgumentMatchers.anyList(), anyBoolean(), anyInt(), anyBoolean(), anyBoolean(),  brokerIdentityCaptor.capture());
+            anyString(), captor.capture(), ArgumentMatchers.anyList(), anyBoolean(), anyInt(), anyBoolean(), anyBoolean(), anyLong(), brokerIdentityCaptor.capture());
         TopicConfigSerializeWrapper wrapper = captor.getValue();
 
         for (TopicConfig topicConfig : topicConfigList) {
